+++ mach_nix/fixes.nix
@@ -7,7 +7,7 @@ let
 
   # evaluate arbitrary python expression and return json parsed result
   py_eval = str: fromJSON ( readFile (
-    pkgs.runCommand "py-eval-result" { buildInputs = with pkgs; [ python python3Packages.packaging ]; } ''
+    pkgs.runCommand "py-eval-result" { buildInputs = with pkgs; [ python3 python3Packages.packaging ]; } ''
       ${pkgs.python}/bin/python -c "${str}" > $out
     ''));
 
@@ -49,6 +49,21 @@ rec {
 #   - ver   (version of the package)         #
 #   - pyver (python version used)            #
 ##############################################
+
+  cartopy.add-native-inputs = {
+    _cond = { prov, ver, ... }: prov == "nixpkgs";
+    nativeBuildInputs.add = with pkgs; [ geos ];
+  };
+
+  cryptography.no-rust-build = {
+    _cond = { prov, ver, ... }: prov == "sdist" && comp_ver ver "<" "3.4";
+    nativeBuildInputs.mod = old: filter (inp: (inp.name or "") != "cargo-setup-hook.sh") old;
+  };
+
+  # remove if merged: https://github.com/NixOS/nixpkgs/pull/114384
+  google-auth.six-input-missing = {
+    propagatedBuildInputs.mod = pySelf: _: oldVal: oldVal ++ [ pySelf.six ];
+  };
 
   httpx.remove-patches = {
     _cond = { prov, ver, ... }: