+++ mach_nix/nix/lib.nix
@@ -170,7 +170,8 @@ rec {
   # Returns `overrides` and `select_pkgs` which satisfy your requirements
   compileOverrides = args:
     let
-      result = import "${compileExpression args}/share/mach_nix_file.nix" { inherit (args) pkgs python; };
+      file = "${compileExpression args}/share/mach_nix_file.nix";
+      result = import file { inherit (args) pkgs python; };
       manylinux =
         if args.pkgs.stdenv.hostPlatform.system == "x86_64-darwin" then
           []
@@ -179,6 +180,7 @@ rec {
     in {
       overrides = result.overrides manylinux autoPatchelfHook;
       select_pkgs = result.select_pkgs;
+      expr = readFile file;
     };
 
   meets_cond = oa: condition:
@@ -316,7 +318,6 @@ rec {
     let
       moved = {
         pkgs = "pkgs";
-        python = "python";
         pypi_deps_db_commit = "pypiDataRev";
         pypi_deps_db_sha256 = "pypiDataSha256";
       };